Refine word-choice from 'reminders' to 'warnings'
While the intent to refer to the messages as 'reminders' during the
Recommendation Phase is to use softer language to match the spirit
of the initial soft-rollout, it's more accurate to refer to the
messages as 'warnings'.

The messages themselves are even prefixed with the word 'Warning',
so it seems fitting to just call it like it is.