Change slug to `making-packages-more-secure`
The dominant convention for existing rubygems.org blog posts, such as it
is, appears to be to slugify and lightly abbreviate the actual post
title, e.g.:

* “March 2022 RubyGems Updates” -> `march-rubygems-updates`
* “API key logging leak” -> `api-key-leak`
* “Removing SHA1 passwords from RubyGems.org” -> `removing-sha1-passwords-in-rubygems-org`
* “Announcing RubyGems.org Stats” -> `announcing-rubygems-stats`
* “RubyGems.org and Chef Gem Ownership” -> `chef-ownership`

So in that spirit — for the sake of consistency alone — I’ve boiled down
the fully descriptive “Making popular Ruby packages more secure” title
to the slightly pithier `making-packages-more-secure`. I think this
preserves enough context (in combination with the rubygems.org domain
at the start of its URL) to accurately summarise the change without
making the URL longer than necessary.